pyproject.toml
[project]
name = "oci-ai"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "langchain-oci>=0.2.1",
    "oci-openai>=1.0.0",
    "openai>=2.14.0",
    "pdfplumber>=0.11.9",
    "python-dotenv>=1.2.1",
]
